refactor(logic): split full_ai_civilian_work_selection into part-of concern files (#3290)
Decompose the 894-line ai/full_ai_civilian_work_selection.dart god-file (the
worst remaining >500-line Phase 0 offender after world_market_phase) into three
additional `part of` concern fragments, matching the established pattern of the
already-merged #3293 / #3297 / #3299 splits and the existing `_feedstock`
fragment:

- full_ai_civilian_work_selection_explore_prospect.dart — mineral exposure
  balancing plus explore / prospect candidate scoring and best-row / combined
  explorer-candidate selection.
- full_ai_civilian_work_selection_build_purchase.dart — Builder
  (build_improvement) and Merchant (purchase_land) candidate scoring / row
  selection and the Old World feedstock unit reservation.
- full_ai_civilian_work_selection_unit_paths.dart — per-unit Builder / Merchant
  / Explorer / lexicographic path appenders and the per-unit dispatcher.

The library file keeps the shared imports, the new `part` directives, the public
result types (FullAiCivilianWorkIdle / FullAiCivilianWorkSelectionResult), the
shared lexicographic / capability helpers, and the public entry point
selectFullAiCivilianWorkOrders (894 -> ~120 lines). Pure structural extraction:
because the fragments are `part of` the same library, every symbol, its
visibility, and private-helper sharing are unchanged, and the public export
surface is untouched, so no consumer imports change. No SPEC change is needed
(epic scope: structural extraction only).

Refs #3290
